fix(stdlib): root STDIN_END_CALLBACKS and split readline tests (#8861)
